+++ compas_python_utils/preprocessing/runSubmit.py
@@ -76,8 +76,8 @@ class pythonProgramOptions:
 
         if (compas_logs_output_override is None):
             self.stringChoices['--output-path'] = os.getcwd()
-            self.stringChoices[
-                '--output-container'] = output_directory  # names the directory to be created and in which log files are created.  Default in COMPAS is "COMPAS_Output"
+            # names the directory to be created and in which log files are created.  Default in COMPAS is "COMPAS_Output"
+            self.stringChoices['--output-container'] = output_directory  
         else:
             self.stringChoices['--output-path'] = compas_logs_output_override
             self.stringChoices['--output-container'] = output_directory